Increase the number of lines to display in fzy-tmux
fzy-tmux splits out a tmux pane and runs fzy in it. The size of the
resulting pane, however, cannot be known it advance due to tmux' dynamic
layouting algorithm.
Running fzy with only ten lines of visible results is unnecessarily
limiting in the majority of cases, as the pane's height is likely
bigger. Ideally fzy would support reading the current window size and
adjust the number of output lines dynamically, but for lack of such a
feature, this change just uses a higher number of lines by default.
Under the assumption of a display covering 100 lines (which is probably
an over estimate), a split would make 50 lines available to fzy, and so
this is the value we pick.